+++ settings.c
@@ -73,7 +73,15 @@ PROGMEM const settings_t defaults = {
     .junction_deviation = DEFAULT_JUNCTION_DEVIATION,
     .arc_tolerance = DEFAULT_ARC_TOLERANCE,
     .g73_retract = DEFAULT_G73_RETRACT,
-
+#ifdef BLOCK_BUFFER_DYNAMIC
+  #ifdef PLANNER_BUFFER_BLOCKS
+    .planner_buffer_blocks = PLANNER_BUFFER_BLOCKS,
+  #elif defined(BLOCK_BUFFER_SIZE)
+    .planner_buffer_blocks = BLOCK_BUFFER_SIZE,
+  #else
+    .planner_buffer_blocks = 35,
+  #endif
+#endif
     .flags.legacy_rt_commands = DEFAULT_LEGACY_RTCOMMANDS,
     .flags.report_inches = DEFAULT_REPORT_INCHES,
     .flags.sleep_enable = DEFAULT_SLEEP_ENABLE,
@@ -318,20 +326,32 @@ PROGMEM static const setting_group_detail_t setting_group_detail [] = {
      { Group_Axis, Group_XAxis, "X-axis"},
      { Group_Axis, Group_YAxis, "Y-axis"},
      { Group_Axis, Group_ZAxis, "Z-axis"},
-#ifdef A_AXIS
+#ifndef AXIS_REMAP_ABC2UVW
+  #ifdef A_AXIS
      { Group_Axis, Group_AAxis, "A-axis"},
-#endif
-#ifdef B_AXIS
+  #endif
+  #ifdef B_AXIS
      { Group_Axis, Group_BAxis, "B-axis"},
-#endif
-#ifdef C_AXIS
+  #endif
+  #ifdef C_AXIS
      { Group_Axis, Group_CAxis, "C-axis"},
-#endif
-#ifdef U_AXIS
+  #endif
+  #ifdef U_AXIS
      { Group_Axis, Group_UAxis, "U-axis"},
-#endif
-#ifdef V_AXIS
+  #endif
+  #ifdef V_AXIS
      { Group_Axis, Group_VAxis, "V-axis"}
+  #endif
+#else
+  #ifdef A_AXIS
+     { Group_Axis, Group_AAxis, "U-axis"},
+  #endif
+  #ifdef B_AXIS
+     { Group_Axis, Group_BAxis, "V-axis"},
+  #endif
+  #ifdef C_AXIS
+     { Group_Axis, Group_CAxis, "W-axis"},
+  #endif
 #endif
 };
 
@@ -518,12 +538,26 @@ PROGMEM static const setting_detail_t setting_detail[] = {
 #if COMPATIBILITY_LEVEL <= 1
      { Setting_DisableG92Persistence, Group_General, "Disable G92 persistence", NULL, Format_Bool, NULL, NULL, NULL, Setting_IsExtendedFn, set_g92_disable_persistence, get_int, NULL },
 #endif
-#if N_AXIS == 4
-     { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "A-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
-#elif N_AXIS == 5
-     { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "A-Axis,B-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
-#elif N_AXIS > 5
-     { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "A-Axis,B-Axis,C-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+#ifndef AXIS_REMAP_ABC2UVW
+  #if N_AXIS == 4
+   { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "A-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+  #elif N_AXIS == 5
+   { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "A-Axis,B-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+  #elif N_AXIS == 6
+   { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "A-Axis,B-Axis,C-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+  #elif N_AXIS == 7
+   { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "A-Axis,B-Axis,C-Axis,U-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+  #elif N_AXIS == 8
+   { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "A-Axis,B-Axis,C-Axis,U-Axis,V-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+  #endif
+#else
+  #if N_AXIS == 4
+     { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "U-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+  #elif N_AXIS == 5
+     { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "U-Axis,V-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+  #elif N_AXIS = 6
+     { Settings_Axis_Rotational, Group_Stepper, "Rotational axes", NULL, Format_Bitfield, "U-Axis,V-Axis,W-Axis", NULL, NULL, Setting_IsExtendedFn, set_rotational_axes, get_int, NULL },
+  #endif
 #endif
 #ifndef NO_SAFETY_DOOR_SUPPORT
      { Setting_DoorSpindleOnDelay, Group_SafetyDoor, "Spindle on delay", "s", Format_Decimal, "#0.0", "0.5", "20", Setting_IsExtended, &settings.safety_door.spindle_on_delay, NULL, is_setting_available },
@@ -531,6 +565,9 @@ PROGMEM static const setting_detail_t setting_detail[] = {
 #endif
      { Setting_SpindleOnDelay, Group_Spindle, "Spindle on delay", "s", Format_Decimal, "#0.0", "0.5", "20", Setting_IsExtended, &settings.safety_door.spindle_on_delay, NULL, is_setting_available },
      { Setting_SpindleType, Group_Spindle, "Default spindle", NULL, Format_RadioButtons, spindle_types, NULL, NULL, Setting_IsExtendedFn, set_spindle_type, get_int, is_setting_available },
+#ifdef BLOCK_BUFFER_DYNAMIC
+     { Setting_PlannerBlocks, Group_General, "Planner buffer blocks", NULL, Format_Int16, "####0", "30", "1000", Setting_IsExtended, &settings.planner_buffer_blocks, NULL, NULL, true },
+#endif
 };
 
 #ifndef NO_SETTINGS_DESCRIPTIONS
@@ -679,6 +716,9 @@ PROGMEM static const setting_descr_t setting_descr[] = {
     { Setting_DoorSpindleOnDelay, "Delay to allow spindle to spin up when spindle at speed tolerance is > 0." },
 #endif
     { Setting_SpindleType, "Spindle selected on startup." },
+#ifdef BLOCK_BUFFER_DYNAMIC
+    { Setting_PlannerBlocks, "Number of blocks in the planner buffer." },
+#endif
 };
 
 #endif
@@ -1794,6 +1834,11 @@ bool read_global_settings ()
     if(settings.spindle.flags.type >= spindle_get_count())
         settings.spindle.flags.type = 0;
 
+#ifdef BLOCK_BUFFER_DYNAMIC
+    if(settings.planner_buffer_blocks < 30 || settings.planner_buffer_blocks > 1000)
+        settings.planner_buffer_blocks = 35;
+#endif
+
     sys.mode = settings.mode;
 
     if(!(hal.driver_cap.spindle_sync || hal.driver_cap.spindle_pid))
@@ -2176,6 +2221,11 @@ setting_datatype_t setting_datatype_to_external (setting_datatype_t datatype)
 bool setting_is_list (const setting_detail_t *setting)
 {
     return setting->datatype == Format_Bitfield || setting->datatype == Format_XBitfield || setting->datatype == Format_RadioButtons;
+}
+
+bool setting_is_integer (const setting_detail_t *setting)
+{
+    return setting->datatype == Format_Integer || setting->datatype == Format_Int8 || setting->datatype == Format_Int16;
 }
 
 static char *remove_element (char *s, uint_fast8_t entry)